requirements.txt
scrapy
mysql-connector-python
flask
python-dotenv